[ios] Fix all warnings (#4738)
# Why

Fixes https://github.com/expo/expo/issues/4429.

# How

- added `Bolts` to `Podfile` to be able to suppress warnings from this specific pod,
- added setting deployment target for all development pods
- went through every warning and fixed

# Test Plan

The project compiles, Camera works, ImagePicker works ok.

+++ ios/Exponent/Kernel/Api/EXApiV2Result.m
@@ -26,7 +26,7 @@ NS_ASSUME_NONNULL_BEGIN
 - (NSString *)description
 {
   return [NSString stringWithFormat:@"<%@: %p, httpStatusCode: %ld, data: %@, error: %@>",
-          NSStringFromClass([self class]), self, _httpStatusCode, _data, _error];
+          NSStringFromClass([self class]), self, (long) _httpStatusCode, _data, _error];
 }
 
 @end
